% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{associate_delegate_to_resource}
\alias{associate_delegate_to_resource}
\title{Adds a member to the resource's set of delegates}
\usage{
associate_delegate_to_resource(
  OrganizationId = "string",
  ResourceId = "string",
  EntityId = "string"
)
}
\arguments{
\item{OrganizationId}{[required] The organization under which the resource exists.}

\item{ResourceId}{[required] The resource for which members are associated.}

\item{EntityId}{[required] The member (user or group) to associate to the resource.}
}
\description{
Adds a member to the resource's set of delegates.
}
\examples{

}
