% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{create_alias}
\alias{create_alias}
\title{Adds an alias to the set of a given member of Amazon WorkMail}
\usage{
create_alias(OrganizationId, EntityId, Alias)
}
\arguments{
\item{OrganizationId}{The organization under which the member exists.}

\item{EntityId}{The alias is added to this Amazon WorkMail entity.}

\item{Alias}{The alias to add to the user.}
}
\description{
Adds an alias to the set of a given member of Amazon WorkMail.
}
\examples{

}
