% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{delete_group}
\alias{delete_group}
\title{Deletes a group from Amazon WorkMail}
\usage{
delete_group(
  OrganizationId = "string",
  GroupId = "string"
)
}
\arguments{
\item{OrganizationId}{[required] The organization that contains the group.}

\item{GroupId}{[required] The identifier of the group to be deleted.}
}
\description{
Deletes a group from Amazon WorkMail.
}
\examples{

}
