% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{delete_resource}
\alias{delete_resource}
\title{Deletes the specified resource}
\usage{
delete_resource(OrganizationId, ResourceId)
}
\arguments{
\item{OrganizationId}{The identifier associated with the organization for which the resource is deleted.}

\item{ResourceId}{The identifier of the resource to be deleted.}
}
\description{
Deletes the specified resource.
}
\examples{

}
