% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{delete_user}
\alias{delete_user}
\title{Deletes a user from Amazon WorkMail and all subsequent systems}
\usage{
delete_user(OrganizationId, UserId)
}
\arguments{
\item{OrganizationId}{[required] The organization that contains the user.}

\item{UserId}{[required] The identifier of the user to be deleted.}
}
\description{
Deletes a user from Amazon WorkMail and all subsequent systems. The action can't be undone. The mailbox is kept as-is for a minimum of 30 days, without any means to restore it.
}
\section{Accepted Parameters}{
\preformatted{delete_user(
  OrganizationId = "string",
  UserId = "string"
)
}
}

