% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{deregister_from_work_mail}
\alias{deregister_from_work_mail}
\title{Mark a user, group, or resource as no longer used in Amazon WorkMail}
\usage{
deregister_from_work_mail(
  OrganizationId = "string",
  EntityId = "string"
)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the Amazon WorkMail entity exists.}

\item{EntityId}{[required] The identifier for the entity to be updated.}
}
\description{
Mark a user, group, or resource as no longer used in Amazon WorkMail. This action disassociates the mailbox and schedules it for clean-up. Amazon WorkMail keeps mailboxes for 30 days before they are permanently removed. The functionality in the console is \emph{Disable}.
}
\examples{

}
