% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{describe_group}
\alias{describe_group}
\title{Returns the data available for the group}
\usage{
describe_group(
  OrganizationId = "string",
  GroupId = "string"
)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the group exists.}

\item{GroupId}{[required] The identifier for the group to be described.}
}
\description{
Returns the data available for the group.
}
\examples{

}
