% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{describe_organization}
\alias{describe_organization}
\title{Provides more information regarding a given organization based on its identifier}
\usage{
describe_organization(OrganizationId)
}
\arguments{
\item{OrganizationId}{The identifier for the organization to be described.}
}
\description{
Provides more information regarding a given organization based on its identifier.
}
\examples{

}
