% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{describe_resource}
\alias{describe_resource}
\title{Returns the data available for the resource}
\usage{
describe_resource(OrganizationId, ResourceId)
}
\arguments{
\item{OrganizationId}{[required] The identifier associated with the organization for which the resource is described.}

\item{ResourceId}{[required] The identifier of the resource to be described.}
}
\description{
Returns the data available for the resource.
}
\section{Accepted Parameters}{
\preformatted{describe_resource(
  OrganizationId = "string",
  ResourceId = "string"
)
}
}

