% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{describe_user}
\alias{describe_user}
\title{Provides information regarding the user}
\usage{
describe_user(OrganizationId, UserId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the user exists.}

\item{UserId}{[required] The identifier for the user to be described.}
}
\description{
Provides information regarding the user.
}
\section{Accepted Parameters}{
\preformatted{describe_user(
  OrganizationId = "string",
  UserId = "string"
)
}
}

