% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{disassociate_delegate_from_resource}
\alias{disassociate_delegate_from_resource}
\title{Removes a member from the resource's set of delegates}
\usage{
disassociate_delegate_from_resource(OrganizationId, ResourceId, EntityId)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the resource exists.}

\item{ResourceId}{[required] The identifier of the resource from which delegates' set members are removed.}

\item{EntityId}{[required] The identifier for the member (user, group) to be removed from the resource's delegates.}
}
\description{
Removes a member from the resource's set of delegates.
}
\section{Accepted Parameters}{
\preformatted{disassociate_delegate_from_resource(
  OrganizationId = "string",
  ResourceId = "string",
  EntityId = "string"
)
}
}

