% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{list_aliases}
\alias{list_aliases}
\title{Creates a paginated call to list the aliases associated with a given entity}
\usage{
list_aliases(
  OrganizationId = "string",
  EntityId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{OrganizationId}{[required] The identifier for the organization under which the entity exists.}

\item{EntityId}{[required] The identifier for the entity for which to list the aliases.}

\item{NextToken}{The token to use to retrieve the next page of results. The first call does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Creates a paginated call to list the aliases associated with a given entity.
}
\examples{

}
