% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{list_groups}
\alias{list_groups}
\title{Returns summaries of the organization's groups}
\usage{
list_groups(OrganizationId, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{OrganizationId}{The identifier for the organization under which the groups exist.}

\item{NextToken}{The token to use to retrieve the next page of results. The first call does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Returns summaries of the organization's groups.
}
\examples{

}
