% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{list_mailbox_permissions}
\alias{list_mailbox_permissions}
\title{Lists the mailbox permissions associated with a mailbox}
\usage{
list_mailbox_permissions(
  OrganizationId = "string",
  EntityId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{OrganizationId}{[required] The identifier of the organization under which the entity (user or group) exists.}

\item{EntityId}{[required] The identifier of the entity (user or group) for which to list mailbox permissions.}

\item{NextToken}{The token to use to retrieve the next page of results. The first call does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Lists the mailbox permissions associated with a mailbox.
}
\examples{

}
