% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{list_organizations}
\alias{list_organizations}
\title{Returns summaries of the customer's non-deleted organizations}
\usage{
list_organizations(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The token to use to retrieve the next page of results. The first call does not contain any tokens.}

\item{MaxResults}{The maximum number of results to return in a single call.}
}
\description{
Returns summaries of the customer's non-deleted organizations.
}
\examples{

}
