% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_package.R
\docType{package}
\name{paws.workmail-package}
\alias{paws.workmail}
\alias{paws.workmail-package}
\title{paws.workmail: Amazon WorkMail}
\description{
Amazon WorkMail is a secure, managed business email
and calendaring service with support for existing desktop and mobile
email clients. You can access your email, contacts, and calendars
using Microsoft Outlook, your browser, or their native iOS and Android
email applications. You can integrate Amazon WorkMail with your
existing corporate directory and control both the keys that encrypt
your data and the location in which your data is stored.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
