% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{update_resource}
\alias{update_resource}
\title{Updates data for the resource}
\usage{
update_resource(OrganizationId, ResourceId, Name = NULL,
  BookingOptions = NULL)
}
\arguments{
\item{OrganizationId}{[required] The identifier associated with the organization for which the resource is updated.}

\item{ResourceId}{[required] The identifier of the resource to be updated.}

\item{Name}{The name of the resource to be updated.}

\item{BookingOptions}{The resource's booking options to be updated.}
}
\description{
Updates data for the resource. It must be preceded by a describe call in order to have the latest information. The dataset in the request should be the one expected when performing another describe call.
}
\section{Accepted Parameters}{
\preformatted{update_resource(
  OrganizationId = "string",
  ResourceId = "string",
  Name = "string",
  BookingOptions = list(
    AutoAcceptRequests = TRUE|FALSE,
    AutoDeclineRecurringRequests = TRUE|FALSE,
    AutoDeclineConflictingRequests = TRUE|FALSE
  )
)
}
}

