% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{associate_delegate_to_resource}
\alias{associate_delegate_to_resource}
\title{Adds a member to the resource's set of delegates}
\usage{
associate_delegate_to_resource(OrganizationId, ResourceId, EntityId)
}
\arguments{
\item{OrganizationId}{The organization under which the resource exists.}

\item{ResourceId}{The resource for which members are associated.}

\item{EntityId}{The member (user or group) to associate to the resource.}
}
\description{
Adds a member to the resource's set of delegates.
}
\examples{

}
