% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{associate_member_to_group}
\alias{associate_member_to_group}
\title{Adds a member to the group's set}
\usage{
associate_member_to_group(
  OrganizationId = "string",
  GroupId = "string",
  MemberId = "string"
)
}
\arguments{
\item{OrganizationId}{[required] The organization under which the group exists.}

\item{GroupId}{[required] The group for which the member is associated.}

\item{MemberId}{[required] The member to associate to the group.}
}
\description{
Adds a member to the group's set.
}
\examples{

}
