% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{create_resource}
\alias{create_resource}
\title{Creates a new Amazon WorkMail resource}
\usage{
create_resource(OrganizationId, Name, Type)
}
\arguments{
\item{OrganizationId}{The identifier associated with the organization for which the resource is created.}

\item{Name}{The name of the created resource.}

\item{Type}{The type of the created resource.}
}
\description{
Creates a new Amazon WorkMail resource. The available types are equipment and room.
}
\examples{

}
