% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{create_user}
\alias{create_user}
\title{Creates a user who can be used in Amazon WorkMail by calling the RegisterToWorkMail operation}
\usage{
create_user(OrganizationId, Name, DisplayName, Password)
}
\arguments{
\item{OrganizationId}{The identifier of the organization for which the user is created.}

\item{Name}{The name for the user to be created.}

\item{DisplayName}{The display name for the user to be created.}

\item{Password}{The password for the user to be created.}
}
\description{
Creates a user who can be used in Amazon WorkMail by calling the RegisterToWorkMail operation.
}
\examples{

}
