% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{delete_mailbox_permissions}
\alias{delete_mailbox_permissions}
\title{Deletes permissions granted to a user or group}
\usage{
delete_mailbox_permissions(OrganizationId, EntityId, GranteeId)
}
\arguments{
\item{OrganizationId}{The identifier of the organization under which the entity (user or group) exists.}

\item{EntityId}{The identifier of the entity (user or group) for which to delete mailbox permissions.}

\item{GranteeId}{The identifier of the entity (user or group) for which to delete granted permissions.}
}
\description{
Deletes permissions granted to a user or group.
}
\examples{

}
