% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.workmail_operations.R
\name{reset_password}
\alias{reset_password}
\title{Allows the administrator to reset the password for a user}
\usage{
reset_password(
  OrganizationId = "string",
  UserId = "string",
  Password = "string"
)
}
\arguments{
\item{OrganizationId}{[required] The identifier of the organization that contains the user for which the password is reset.}

\item{UserId}{[required] The identifier of the user for whom the password is reset.}

\item{Password}{[required] The new password for the user.}
}
\description{
Allows the administrator to reset the password for a user.
}
\examples{

}
